Add 3pin hardware serial
All terminal functions and commands are now accessible either from the
STM32F103 USB serial port or from PA2 and PA3 via an external USB
adapter such as an FTDI 3.3V adapter.

This enhancement was to work around the RPI not detecvting the STM32
USBSerial and the USBSerial not being accessible from a PC through a
USB3.0 hub.